pyproject.toml
[project]
name = "backdoor-leakage"
version = "0.1.0"
description = "Backdoor Leakage"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "llmcompare",
    "pandas>=2.3.1",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "pytest",
    "ruff",
    "mypy",
]

[tool.uv]
default-groups = "all"

[tool.uv.sources]
llmcompare = { git = "https://github.com/johny-b/llmcompare.git", rev = "main" }
